requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1

# LlamaIndex core
llama-index==0.11.20
llama-index-core==0.11.20
llama-index-embeddings-huggingface==0.3.1
llama-index-vector-stores-chroma==0.2.2
llama-index-llms-openai==0.2.16

# Chroma vector store
chromadb==0.5.3

# Hugging Face Inference API client
huggingface_hub==0.25.1

# LiteLLM for OpenRouter LLM routing
litellm==1.48.0